+++ anusha_chettukindi/c logics/min_max_diff and pick count sum matched with length of array.c
